"split" feauture was incorrectly ordering lines
Although it did not show on the screen, whenever the "split" button was clicked, internally the lines were being put in the wrong order. The result was that, even though it all looked fine on the screen, when you saved the data, the resulting json file would be all messed up.